backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1

# LangChain
langchain==0.2.1
langchain-core==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.4

# LLM / Embeddings
groq==0.8.0
sentence-transformers==2.7.0

# Vector Store
chromadb>=0.4.0

# Database
aiosqlite==0.20.0
